// src/app/main/products/models/product.model.ts
export class ProductModel {
    id: number;
    number: string;
    name: string;
    description: string;
    material_description: string
    brand: any;
    brands_id: number;
    sale_price: number;
    amount: number;
    category: any;
    categorys_id: number;
    url_image: string;
    url_complete: string;
    stock: number;
    unit_symbol: string;
    inventories_id: number;
}